+++ .eslintrc.js
@@ -20,7 +20,7 @@ module.exports = {
             
         ],
         "linebreak-style": [
-            "error",
+            "off",
             "unix"
         ],
         "quotes": [
